Move reducers one level up and hook the up
